+++ checkstyle.xml
@@ -181,7 +181,7 @@
 			- net.fabricmc.fabric.(impl|mixin).<module-name>[.v<version>][.<extra packages...>]
 			- <any legacy package>
 			where <module-name> is a set of '.'-separated words, all in singular (not ending with s except for ss) and not starting with client. or server.
-			and <version> is a positive integer (1, 2, 3, ...)
+			and <version> is a non-negative integer (0, 1, 2, 3, ...)
 			and <extra packages...> is a set of '.'-separated words with all the first potentially containing digits.
 			Negative lookahead ensures that client/server can't be replaced with common disguised as the module name.
 
@@ -194,7 +194,7 @@
 			- largely unconstained trailing subpackages
 			-->
 			<property name="format"
-			value="^net\.fabricmc\.fabric\.(api(?!\.common\.)(\.client|\.server|)(\.(?!client\.|server\.)[a-z]+([a-rt-z]|ss))+\.v[1-9][0-9]*|(impl|mixin|test)(?!\.common\.)(\.client|\.server|)(\.(?!client\.|server\.)[a-z]+([a-rt-z]|ss))+(\.v[1-9][0-9]*)?|api\.(event|util|biomes\.v1|registry|client\.screen|container|block|entity|client\.itemgroup|client\.keybinding|tag|tools|client\.model|network|server|client\.render|resource|client\.texture))(|\.[a-z]+(\.[a-z0-9]+)*)$"/>
+			value="^net\.fabricmc\.fabric\.(api(?!\.common\.)(\.client|\.server|)(\.(?!client\.|server\.)[a-z]+([a-rt-z]|ss))+\.v[0-9]+|(impl|mixin|test)(?!\.common\.)(\.client|\.server|)(\.(?!client\.|server\.)[a-z]+([a-rt-z]|ss))+(\.v[0-9]+)?|api\.(event|util|biomes\.v1|registry|client\.screen|container|block|entity|client\.itemgroup|client\.keybinding|tag|tools|client\.model|network|server|client\.render|resource|client\.texture))(|\.[a-z]+(\.[a-z0-9]+)*)$"/>
 		</module>
 
 		<!--<module name="InvalidJavadocPosition"/>-->
