+++ fabric-lifecycle-events-v1/src/client/java/net/fabricmc/fabric/impl/client/event/lifecycle/ClientLifecycleEventsImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.event.lifecycle;
+package net.fabricmc.fabric.impl.client.event.lifecycle;
 
 import net.minecraft.block.entity.BlockEntity;
 
@@ -23,6 +23,7 @@ import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientBlockEntityEvents;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientChunkEvents;
+import net.fabricmc.fabric.impl.event.lifecycle.LoadedChunksCache;
 
 @Environment(EnvType.CLIENT)
 public final class ClientLifecycleEventsImpl implements ClientModInitializer {
