+++ fabric-networking-api-v1/src/client/java/net/fabricmc/fabric/mixin/networking/client/accessor/MinecraftClientAccessor.java
@@ -14,14 +14,22 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.client.rendering;
+package net.fabricmc.fabric.mixin.networking.client.accessor;
 
-import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.fabric.api.client.rendering.v1.InvalidateRenderStateCallback;
+import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
 
-public class RenderingCallbackInvoker implements ClientModInitializer {
-	@Override
-	public void onInitializeClient() {
-		InvalidateRenderStateCallback.EVENT.register(() -> net.fabricmc.fabric.api.client.render.InvalidateRenderStateCallback.EVENT.invoker().onInvalidate());
-	}
+import net.minecraft.client.MinecraftClient;
+import net.minecraft.network.ClientConnection;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+
+@Environment(EnvType.CLIENT)
+@Mixin(MinecraftClient.class)
+public interface MinecraftClientAccessor {
+	@Nullable
+	@Accessor("integratedServerConnection")
+	ClientConnection getConnection();
 }
